% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vifmiqo.R
\name{vifmiqo}
\alias{vifmiqo}
\title{Eliminate Multicollinearity}
\usage{
vifmiqo(X, y, alpha = 5)
}
\arguments{
\item{X}{Dataframe containing potential covariates and their values}

\item{y}{Dataframe with output values corresponding to X}

\item{alpha}{Optional- standard constraint > 1}
}
\value{
A list containing the optimal subset of parameters ("") and the corresponding
total number selected ("S"), R^2 for linear regression ("Rsq"), and VIF value ("VIF_max")
}
\description{
This function loads data in two pieces, X and y. Then it selects the subset
of variables that minimize multicollinearity using a mixed-integer optimization
framework.
}
